Move distribution types to their own header
