+++ benchmarks/key_generator.hpp
@@ -17,6 +17,7 @@
 #pragma once
 
 #include <defaults.hpp>
+#include <distribution.hpp>
 
 #include <nvbench/nvbench.cuh>
 
@@ -38,19 +39,6 @@
 #include <type_traits>
 
 namespace cuco::benchmark {
-
-namespace dist_type {
-struct unique {
-};
-
-struct uniform {
-  int64_t multiplicity;  // TODO assert >0
-};
-
-struct gaussian {
-  double skew;  // TODO assert >0
-};
-};  // namespace dist_type
 
 /**
  * @brief Random key generator.
@@ -233,26 +221,4 @@ class key_generator {
   RNG rng_;  ///< Random number generator
 };
 
-template <typename Dist>
-auto dist_from_state(nvbench::state const& state)
-{
-  if constexpr (std::is_same_v<Dist, dist_type::unique>) {
-    return Dist{};
-  } else if constexpr (std::is_same_v<Dist, dist_type::uniform>) {
-    auto const multiplicity = state.get_int64_or_default("Multiplicity", defaults::MULTIPLICITY);
-    return Dist{multiplicity};
-  } else if constexpr (std::is_same_v<Dist, dist_type::gaussian>) {
-    auto const skew = state.get_float64_or_default("Skew", defaults::SKEW);
-    return Dist{skew};
-  } else {
-    // TODO static assert fail
-  }
-}
-
 }  // namespace cuco::benchmark
-
-NVBENCH_DECLARE_TYPE_STRINGS(cuco::benchmark::dist_type::unique, "UNIQUE", "dist_type::unique");
-NVBENCH_DECLARE_TYPE_STRINGS(cuco::benchmark::dist_type::uniform, "UNIFORM", "dist_type::uniform");
-NVBENCH_DECLARE_TYPE_STRINGS(cuco::benchmark::dist_type::gaussian,
-                             "GAUSSIAN",
-                             "dist_type::gaussian");
